mainloop: React to expose events when the engine is paused
When the engine is paused, it doesn't redraw the window content anymore.
On uncomposited environments, this means that the window content can be
lost. Enable expose events and react to them when the engine is paused
(if it's not, the whole window will be repainted soon anyway, so we can
ignore them).